Patch_sampling possible w/ mask_level > max_level
When the level of the mask is set to be greater than the number of
levels of the slide (max_level), then get_whole_image immediately adapt
the level to slide.level_count - 1. This commit updates level for the
rest of the procedure in Patch_sampling, otherwise we would get an
error.